refactor: client 패키지명 수정 및 위치 변경, isApiCallSuccessful 메소드명 변경 #9
