+++ src/test/java/com/supernova/ssagrissakssak/feed/controller/BoardControllerTest.java
@@ -1,0 +1,55 @@
+package com.supernova.ssagrissakssak.feed.controller;
+
+import com.supernova.ssagrissakssak.feed.controller.request.UserCreateRequest;
+import com.supernova.ssagrissakssak.feed.service.BoardService;
+import com.supernova.ssagrissakssak.mockuser.MockUser;
+import org.hibernate.annotations.Mutability;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.http.MediaType;
+import org.springframework.restdocs.payload.JsonFieldType;
+import org.springframework.test.web.servlet.MockMvc;
+
+import static org.junit.jupiter.api.Assertions.*;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.BDDMockito.given;
+import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
+import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.post;
+import static org.springframework.restdocs.payload.PayloadDocumentation.*;
+import static org.springframework.restdocs.payload.PayloadDocumentation.fieldWithPath;
+import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+@WebMvcTest(BoardController.class)
+class BoardControllerTest extends RestDocsSupport {
+
+    @Autowired
+    private MockMvc mockMvc;
+
+    @MockBean
+    private BoardService boardService;
+
+    @Test
+    @MockUser
+    void likeBoardContent_Success() throws Exception {
+        // given
+        String contentId = "test123";
+
+        doNothing().when(boardService).likeBoardContent(contentId);
+
+        // when & then
+        mockMvc.perform(post("/boards/{contentId}/like", contentId))
+                .andExpect(status().isOk())
+                .andDo(document("board-like",
+                        pathParameters(
+                                parameterWithName("contentId").description("게시물 ID")
+                        ),
+                        responseFields(
+                                fieldWithPath("status").description("응답 상태 코드"),
+                                fieldWithPath("message").description("응답 메시지")
+                        )
+                ));
+    }
+}
