+++ src/test/java/com/supernova/ssagrissakssak/feed/service/BoardServiceTest.java
@@ -1,0 +1,11 @@
+package com.supernova.ssagrissakssak.feed.service;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+class BoardServiceTest {
+
+    @Test
+    void likeBoardContent() {
+    }
+}
